base for graphics system revamp

+++ include/components/position.hpp
@@ -31,21 +31,21 @@ namespace Components
     {
     private:
         Position(std::shared_ptr<Entity> parent,
-                 float x_,
-                 float y_);
+                 int x_,
+                 int y_);
 
     public:
-        float x, y;
+        int x, y;
 
-        inline Vector2f toVector() const { return Vector2f(x, y); }
+        inline Vector2i toVector() const { return Vector2i(x, y); }
 
         std::string toString() const;
 
         static std::shared_ptr<Position> create(std::shared_ptr<Entity> parent,
-                                                float x,
-                                                float y);
+                                                int x,
+                                                int y);
         static std::shared_ptr<Position> create(std::shared_ptr<Entity> parent,
-                                                const Vector2f &v = Vector2f::ZERO);
+                                                const Vector2i &v = Vector2i::ZERO);
 		static std::shared_ptr<Position> create(std::shared_ptr<Entity> parent,
 												const MapPosition &position);
 	};
